pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loxone-mcp"
version = "0.1.0"
description = "MCP server bridging AI systems with Loxone home automation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
authors = [
    { name = "Loxone MCP Contributors" },
]
keywords = ["mcp", "loxone", "home-automation", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.14",
    "Topic :: Home Automation",
]
dependencies = [
    "mcp>=1.9.0",
    "websockets>=15.0",
    "aiohttp>=3.11.0",
    "cryptography>=44.0.0",
    "prometheus-client>=0.22.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.8.0",
    "structlog>=25.1.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.0",
    "mypy>=1.15.0",
    "pip-audit>=2.9.0",
]

[project.scripts]
loxone-mcp = "loxone_mcp.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

# T003: Ruff configuration
[tool.ruff]
target-version = "py314"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific rules
    "ASYNC", # flake8-async
    "S",     # flake8-bandit (security)
    "LOG",   # flake8-logging
]
ignore = [
    "S101",  # assert usage (needed in tests)
    "S104",  # possible binding to all interfaces
]

[tool.ruff.lint.per-file-ignores]
"src/tests/**/*.py" = ["S101", "S106"]

[tool.ruff.lint.isort]
known-first-party = ["loxone_mcp"]

# T004: Mypy configuration
[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "mcp.*",
    "prometheus_client.*",
    "websockets.*",
    "aiohttp.*",
]
ignore_missing_imports = true

# T005: Pytest configuration
[tool.pytest.ini_options]
testpaths = ["src/tests"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=loxone_mcp",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=80",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (may need network)",
    "contract: MCP protocol contract tests",
    "slow: Tests that take more than 1 second",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["loxone_mcp"]
omit = [
    "*/tests/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "raise NotImplementedError",
    "@overload",
]
